Implement host defined variables support.
It's implemented via pointer indirection: whenever there's a program
scope variable declaration that might be host defined, create a pointer
variable for it instead.  If the variable is later defined in the
translation unit, create the variable definition and assign the pointer
to its address (and make it private), which causes it to be optimized
away back to the variable definition.

In case the variable is not defined within the translation unit,
it is a host defined variable in which case the runtime assigns the
pointer value.