[BRIG] (almost) always scalarize and promote MULT_HIGHPART_EXPR
It seems the support for vector MULT_HIGHPART_EXPR is flaky and
undertested.

The robust thing here is to force scalarization always with MULHI for
now, until these issues are  debugged further. I added an exception for
2x64b MULT_HIGHPART_EXPR to avoid the need for 128b scalar arithmetics,
and as it seems to work for the CPUs I've tested.

The decision should not IMO belong to the frontend, but there'd be
better a step where the vector operations are optionally scalarized if
the target prefers scalarized operations which should be caught by this
step also. Something to fix during optimization work.