Ensure per WI copies of private variables are aligned too.
We need to take care of two cases of alignment with private
variables because of the layout where the same variable for
each work-item is laid out in successive addresses.

1) Ensure the first work-item's variable is in an aligned
offset:

2) Each successive per-work-item copy should be aligned.
If the variable has wider alignment than size then we need
to add extra padding to ensure it.  The padding must be
included in the size to allow per-work-item offset computation
to find their own aligned copy