Provide access to HSA WAVESIZE value
Hi,

when writing HSA instructions directly, WAVESIZE can be used at many
places where an immediate or a register can. That is not easily possible
in a higher level language, of course, but still it might be useful to
get at its value.  This patch provides a new builtin that will load it
to a register.

Thanks,

Martin